แก้ไขการส่ง error messege ใน user resolver

+++ graphql-api/src/graphql/resolvers/users.js
@@ -24,21 +24,17 @@ export default {
     },
     Mutation: {
         createUser: async (parent, { name, email, username, password }) => {
-            try{
-                if(await db('users').select('*').where('username', username).first()){
-                    return "Username already exists";
-                }
-                if(await db('users').select('*').where('email', email).first()){
-                    return "Email already exists";
-                }
-
-                const encryptedPassword = await bcrypt.hash(password, 10);
-
-                await db('users').insert({name: name,email: email, username: username,password: encryptedPassword});
-                return "Create Success";
-            }catch(err){
-                return "Create Fail";
+            if(await db('users').select('*').where('username', username).first()){
+                throw new Error('Username already exists');
             }
+            if(await db('users').select('*').where('email', email).first()){
+                throw new Error('Email already exists');
+            }
+
+            const encryptedPassword = await bcrypt.hash(password, 10);
+
+            await db('users').insert({name: name,email: email, username: username,password: encryptedPassword});
+            return "Create Success";
         },
         changePassword: async (parent, { id, password }) => {
             try{
@@ -46,47 +42,45 @@ export default {
                 await db('users').update({password: encryptedPassword}).where('id', id);
                 return "Change Success";
             }catch(err){
-                return "Change Fail";
+                return err;
             }
         },
         changeImage: async (parent, { id, image }) => {
-            try {
-              const { createReadStream, filename, mimetype, encoding } = await image;
-            //   const stream = createReadStream();
-              const pathName = path.join(process.cwd(), `public/images/users/${filename}`);
-              const writeStream = createWriteStream(pathName);
-          
-              await new Promise((resolve, reject) => {
-                stream.pipe(writeStream)
-                  .on('finish', resolve)
-                  .on('error', reject);
-              });
-          
-              const imagePath = `http://localhost:4000/images/users/${filename}`;
-              await db('users').where('id', id).update({ image: imagePath });
-          
-              return "Change Success";
-            } catch (err) {
-              return "Change Fail";
+            try{
+                const { createReadStream, filename, mimetype, encoding } = await image;
+                const stream = createReadStream();
+                const pathName = path.join(process.cwd(), `public/images/users/${filename}`);
+                const writeStream = createWriteStream(pathName);
+                
+                await new Promise((resolve, reject) => {
+                    stream.pipe(writeStream)
+                    .on('finish', resolve)
+                    .on('error', reject);
+                });
+            
+                const imagePath = `http://localhost:4000/images/users/${filename}`;
+                await db('users').where('id', id).update({ image: imagePath });
+    
+                return "Change Success";
+            }catch(err){
+                return err;
             }
           },
         changeEmail: async (parent, { id, email }) => {
-            try{
-                if(await db('users').select('*').where('email', email).first()){
-                    return "Email already exists";
-                }
-                await db('users').update({email: email}).where('id', id);
-                return "Change Success";
-            }catch(err){
-                return "Change Fail";
+            const emailExist = await db('users').select('*').where('email', email).first();
+            if(emailExist){
+                    throw new Error('Email already exists')
             }
+            await db('users').update({email: email}).where('id', id);
+            return "Change Success";
         },
         deleteUser: async (parent, { id }) => {
-            try{
+            const rows = await db('users').select('*').where('username', username).first();
+            if(rows){
                 await db('users').delete().where('id', id);
                 return "Delete Success";
-            }catch(err){
-                return "Delete Fail";
+            }else{
+                throw new Error('User not found');
             }
         },
         login: async (parent, { username, password }) => {
